Deabstractified the stochastic hmc algos
